Update storj-system-health.sh
added a check of the LOGMIN selected part of the log, if there are time lags larger than 3 mins between a started and finished download of GET_AUDIT; the script will alert you, in case there are time lags. the selection is done by satellite, so a count per satellite will be sent by email to have some more details for further analysis.